4MS-main/requirements.txt
# Core dependencies
streamlit>=1.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
httpx>=0.27.0
pandas>=2.1.0
pytz>=2023.3
pydantic>=2.6.0
tenacity>=8.2.3

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.22

# LangChain and LangGraph dependencies
langchain==0.3.26
langchain_google_genai
langchain-community>=0.0.10
langchain-core
langchain-huggingface
langchain-text-splitters
langgraph
langgraph-sdk

# Google AI/Cloud dependencies
google-genai
google-generativeai
google-api-python-client>=2.140.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# Vector stores
langchain-chroma
chromadb
#pysqlite3-binary

# Document processing
pypdf>=3.15.1
PyPDF2>=3.0.0

# Audio processing
openai-whisper
assemblyai>=0.17.0
audio-recorder-streamlit>=0.0.8
av>=10.0.0  # Required for audio processing
pydub>=0.25.1  # For audio duration detection
websockets>=12.0  # For AssemblyAI real-time streaming
elevenlabs>=1.2.0

edge-tts>=6.1.9  

# PDF generation
fpdf>=1.7.2
reportlab>=4.0.4

# Visualization
graphviz>=0.20.1
ipython>=8.12.0
sentence-transformers

# Fundamental Knowledge Assessment & CV filtering (Groq)
groq>=0.4.0
langchain-groq
openpyxl>=3.1.0
python-docx>=1.0.0
plotly>=5.17.0
